futurize -f lib2to3.fixes.fix_long

+++ lib/python/Components/Network.py
@@ -681,7 +681,7 @@ class Network:
 		mask = 1L<<31
 		xnet = (1L<<32)-1
 		cidr_range = range(0, 32)
-		cidr = long(nmask)
+		cidr = int(nmask)
 		if cidr not in cidr_range:
 			print('cidr invalid: %d' % cidr)
 			return None
