+++ lib/python/Components/NimManager.py
@@ -513,7 +513,7 @@ class SecConfigure:
 					if c.has_key(cdc):
 						sec.setCommittedCommand(c[cdc])
 					else:
-						sec.setCommittedCommand(long(cdc))
+						sec.setCommittedCommand(int(cdc))
 
 					sec.setFastDiSEqC(currLnb.fastDiseqc.value)
